+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/query/AggregationQueryService.java
@@ -22,6 +22,9 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import org.apache.skywalking.oap.server.core.query.type.debugging.DebuggingTrace;
+import org.apache.skywalking.oap.server.core.query.type.debugging.DebuggingSpan;
+import org.apache.skywalking.oap.server.core.query.type.debugging.DebuggingTraceContext;
 import org.apache.skywalking.oap.server.library.util.StringUtil;
 import org.apache.skywalking.oap.server.core.Const;
 import org.apache.skywalking.oap.server.core.analysis.IDManager;
@@ -53,7 +56,7 @@ public class AggregationQueryService implements Service {
         return aggregationQueryDAO;
     }
 
-    public List<SelectedRecord> sortMetrics(TopNCondition condition, Duration duration) throws IOException {
+    private List<SelectedRecord> invokeSortMetrics(TopNCondition condition, Duration duration) throws IOException {
         if (!condition.senseScope()) {
             return Collections.emptyList();
         }
@@ -108,5 +111,22 @@ public class AggregationQueryService implements Service {
             }
         });
         return selectedRecords;
+    }
+
+    public List<SelectedRecord> sortMetrics(TopNCondition condition,
+                                            Duration duration) throws IOException {
+        DebuggingTraceContext traceContext = DebuggingTrace.TRACE_CONTEXT.get();
+        DebuggingSpan span = null;
+        try {
+            if (traceContext != null) {
+                span = traceContext.createSpan("Query Service: sortMetrics");
+                span.setMsg("TopNCondition: " + condition + ", Duration: " + duration);
+            }
+            return invokeSortMetrics(condition, duration);
+        } finally {
+            if (traceContext != null && span != null) {
+                traceContext.stopSpan(span);
+            }
+        }
     }
 }