+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/query/input/TopNCondition.java
@@ -20,6 +20,7 @@ package org.apache.skywalking.oap.server.core.query.input;
 
 import lombok.Getter;
 import lombok.Setter;
+import lombok.ToString;
 import org.apache.skywalking.oap.server.core.query.MetricsMetadataQueryService;
 import org.apache.skywalking.oap.server.core.query.enumeration.MetricsType;
 import org.apache.skywalking.oap.server.core.query.enumeration.Order;
@@ -33,6 +34,7 @@ import org.apache.skywalking.oap.server.core.storage.annotation.ValueColumnMetad
  */
 @Setter
 @Getter
+@ToString
 public class TopNCondition {
     /**
      * Metrics name